+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 composer extension Change Log
 2.0.5 under development
 -----------------------
 
+- Bug #11: `generateCookieValidationKey()` now saves config file only when `cookieValidationKey` was generated (rob006)
 
 
 2.0.4 February 06, 2016