+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 composer extension Change Log
 
 - Enh #10: Added `yii\composer\Installer::postInstall()` method (rob006)
 - Bug #11: `generateCookieValidationKey()` now saves config file only when `cookieValidationKey` was generated (rob006)
+- Enh #12: Added `yii\composer\Installer::copyFiles()` method (rob006)
 
 
 2.0.4 February 06, 2016